feat: overlay-kit 추가 및 테스트 환경 설정 개선\n\n- overlay-kit 패키지를 추가하여 다이얼로그 및 모달 관리 개선\n- 테스트 유틸리티를 리팩터링하여 코드 중복 제거 및 가독성 향상\n- 기존 테스트에서 SnackbarProvider를 setup 유틸리티로 통합하여 일관성 유지
