+++ package.json
@@ -30,6 +30,7 @@
     "express": "^4.19.2",
     "msw": "^2.10.3",
     "notistack": "^3.0.2",
+    "overlay-kit": "^1.8.4",
     "react": "19.1.0",
     "react-dom": "19.1.0"
   },
